Drop zero-copy, close #3834
Motivation:

* it seems to be causing performance issues
* it's more or less dead code as most traffic is encrypted nowadays

Modification:

* drop MultipartFileRegion and related code
* drop DefaultFileRegion usage
* drop gatling.http.ahc.enableZeroCopy config option

Result:

Faster multipart upload and less dead code

+++ gatling-http-client/src/main/java/io/gatling/http/client/body/multipart/impl/MultipartChunkedInput.java
@@ -61,7 +61,7 @@ public class MultipartChunkedInput implements ChunkedInput<ByteBuf> {
   public MultipartChunkedInput(List<PartImpl> parts, long contentLength) {
     this.parts = parts;
     this.contentLength = contentLength;
-    this.chunkSize = contentLength > 0 ? (int) Math.min(contentLength, (long) DEFAULT_CHUNK_SIZE) : DEFAULT_CHUNK_SIZE;
+    this.chunkSize = contentLength > 0 ? (int) Math.min(contentLength, DEFAULT_CHUNK_SIZE) : DEFAULT_CHUNK_SIZE;
   }
 
   @Override
